requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
motor==3.1.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
pydantic==1.10.2
Pygments==2.13.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rich==12.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
